refactor(step-1): extract logger and logAndThrow into src/logging.ts
Splits the logger and logAndThrow out of the grab-bag utils.ts into a
pure, zero-internal-import module at src/logging.ts. This is the first
cycle-breaker in the layered refactor: utils <-> notes and utils <->
bear-urls cycles existed partly because logger lived alongside code
that imported from those modules. A pure logging.ts can be reached
from any layer (tools, operations, infra) without creating cycles.

No behavior change. All importers (main, notes, tags, bear-urls,
database, and utils itself) now pull logger/logAndThrow from
./logging.js instead of ./utils.js.
